Fix dropped sales by fetching overlapping windows
It seems that sale/transfer events are inserted into the stream late
with old timestamps: e.g., the timestamp of a `created` or `bid_entered`
event that's first available at time `t` will be approximately `t`, but
the timestamp of a `successful` or `transfer` event that's first
available at time `t` will be the time of its associated transaction,
which may precede `t` by 20 seconds or so. Thus, polling every 5 seconds
for the intervals in those last 5 seconds will not actually capture the
sale/transfer events, because by the time they're available to be
queried, the `occurred_after` filter excludes them.

This patch attempts to work around this issue by expanding the timestamp
window in the query from 5 seconds to 65 seconds, while still polling
every 5 seconds. Visually, we used to poll like this (horizontal axis is
time, vertical axis is queries):

    |<- Q1 ->|
             |<- Q2 ->|
                      |<- Q3 ->|

...but now we poll like this (not to scale):

    |<------- Q1 ------->|
             |<------- Q2 ------->|
                      |<------- Q3 ------->|

Events that occur multiple times are deduplicated on the client side.

Logs show that this suffices to show sales and transfers:

```
[679788595] 2021-08-29T05:34:00.730213 offer_entered    bid 2.533 ETH on   39000377 "Synapses #377"
[679788604] 2021-08-29T05:34:00.822682 transfer         tx 0x8879012997b8279f91475ddb891ebf70987b720e641b7eb88c90e111587b9f1c on   64000299 "AlgoRhythms #299"
[679788697] 2021-08-29T05:34:02.103291 successful       paid 9.400 ETH on   64000299 "AlgoRhythms #299"
[679788724] 2021-08-29T05:34:02.417828 offer_entered    bid 2.740 ETH on   39000306 "Synapses #306"
```