Replace OpenSea event `sources` with full `config`
The set of event sources is now determined from the config object, and
re-computed on each poll loop. This way, if the config reloads, the
event fetching can update its queries appropriately; before this patch,
the `sources` value was set only once at program start.

Small caveat: we continue to track a single `lastEventIds` set, so given
the following sequence:

 1. Initial state: Collection C is in `config.collectionSlugs()`.
 2. We poll for events, find an event E for Collection C, and record it
    in `lastEventIds`.
 3. Collection C is removed from `config.collectionSlugs()`.
 4. We poll again. There are naturally no events for Collection C, so
    event E is removed from `lastEventIds`.
 5. Collection C is added to `config.collectionSlugs()` again.
 6. We poll again. We find event E again. Because it's no longer in
    `lastEventIds`, we pass it to the `handleEvent` callback, even
    though it's been seen before.

We could handle this by keeping a separate `lastEventIds` map for each
collection, at the cost of leaking that memory when a collection is
removed but never added again. It seems like not a big deal either way,
so I'm keeping the code simpler for now.

case of the follo
collection is removed from the sources